refactor: organized code, fixed double calls...
When deleting a scream, likes, comments and notifications are deleted.
Deleting a like triggers the onDelete firestore listener on the likes
collection, which would try and delete the corresponding notification.
This little execution duplication has been removed by not deleting
notifications with the scream, but by letting the onDelete of the
corresponding collection delete the related notification.